Squashed commit:
[4b21748] #70 added make_ab_plane_boundary_data

[06bd50b] #70 added a sample code using cielab.py.
